settings: Slims new Power Menu (2/2)
gmillz already prepared it and I finished it now. Idea was to be more
flexible on the power menu but at the same time not complicated.
So we adapted Power Menu to our dslv system which allows to sort
drag swipe etc the entries.

So user can now have the power menu how they like it.
As well custom apps or activities can be added

So currently:
airplanemode
custom apps
screenshot
reboot
sound
power off

are available as options to add or remove or customize with our
easy dslv settings

As well we are only using our central classes for it....means adding
whatever shortcut is mostly only a 10 line change in future

Change-Id: I8a3f99bdd0ffeab84b8b4067562d81f09c862273

Conflicts:
	res/values/strings.xml
	res/xml/power_menu_fragment.xml
	res/xml/power_menu_settings.xml
	res/xml/preference_headers.xml